Strip unreproduced defensive code and back the confidence cap with a repro test
Reverts the signed sequence-number delta in FindGaps (LIS-reconstructed paths
are co-monotonic in time and sequence, and on .NET Core the concatenator's
old cast saturated rather than wrapped - no reversal was ever reproduced) and
drops the non-positive denominator guard in ConfidenceCalculator (requires
non-positive track length metadata nothing produces). The confidence cap
itself is now covered by a regression test constructing a coverage whose
matched path overhangs the declared track length: pre-cap it yielded ~1.11.

+++ src/SoundFingerprinting/Query/MatchedWithExtensions.cs
@@ -447,10 +447,7 @@ namespace SoundFingerprinting.Query
                 float startsAt = matches[i - 1].Item2;
                 float endsAt = matches[i].Item2;
                 float gap = (float)SubFingerprintsToSeconds.GapLengthToSeconds(endsAt, startsAt, fingerprintLength);
-                // overflow hygiene, not a reproduced defect: LIS-reconstructed paths are co-monotonic in time and
-                // sequence numbers, but BestPath is not validated when stitched from realtime entries or
-                // deserialized, and an unsigned delta would wrap on a reversal instead of comparing sanely
-                bool sequenceNumberIncremented = (long)matches[i].Item1 - matches[i - 1].Item1 > 1;
+                bool sequenceNumberIncremented = matches[i].Item1 - matches[i - 1].Item1 > 1;
                 float start = endsAt - gap;
                 if (!(endsAt <= start) && gap > permittedGap && sequenceNumberIncremented)
                 {
